#100 modelとserializeとviewをディレクトリ化する
