+++ api/app/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Product, AccessoryCategory, MaterialCategory, SalesSite, ProductImage, ProducerProfile, SkillMarket, Sns
+from app.models import Product, AccessoryCategory, MaterialCategory, SalesSite, ProductImage, ProducerProfile, SkillMarket, Sns
 
 admin_model_list = [
     Product,
