+++ api/app/serializers/producer_profile.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from app.models.producer_profile import ProducerProfile
+
+
+class ProducerProfileSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = ProducerProfile
+        fields = ('name', 'introduction', 'logo')
