+++ api/app/views/__init__.py
@@ -1,0 +1,4 @@
+from app.views.category import AccessoryCategoryListAPIView, MaterialCategoryListAPIView
+from app.views.site import SalesSiteListAPIView, SkillMarketListAPIView, SnsListAPIView
+from app.views.product import ProductListAPIView, ProductRetrieveAPIView
+from app.views.producer_profile import ProducerProfileAPIView
